updated almost finish

+++ routes/web.php
@@ -1,6 +1,7 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\AuthController;
 use App\Http\Controllers\ClassController;
 use App\Http\Controllers\StudentController;
 use App\Http\Controllers\TeacherController;
@@ -21,35 +22,35 @@ Route::get('/train', function() {
 });
 Route::get('/', function () {
     return view('welcome');
-});
+})->middleware('auth');
 
-Route::get('/login', function () {
-    return view('login');
-});
+Route::get('/login', [AuthController::class, 'login'])->name('login')->middleware('guest');
+Route::post('/login-fill', [AuthController::class, 'loginFill'])->middleware('guest');
 
-Route::get('/register', function () {
-    return view('register');
-});
+Route::get('/register', [AuthController::class, 'register'])->name('register')->middleware('guest');
+Route::post('/register-fill', [AuthController::class, 'registerFill'])->middleware('guest');
 
-Route::get('/student', [StudentController::class, 'student']);
-Route::get('/student/{id}', [StudentController::class, 'details']);
-Route::get('/student-add', [StudentController::class, 'create']);
-Route::post('/students', [StudentController::class, 'insert']);
-Route::put('student-edit/{id}', [StudentController::class, 'edit']);
-Route::get('student-del/{id}', [StudentController::class, 'del']);
-Route::delete('student-delete/{id}', [StudentController::class, 'delete']);
-Route::get('/student-deleted', [StudentController::class, 'deleted']);
-Route::get('/student/{id}/restore', [StudentController::class, 'restore']);
+Route::get('/logout', [AuthController::class, 'logout'])->middleware('auth');
+
+Route::get('/student', [StudentController::class, 'student'])->middleware('auth');
+Route::get('/student/{id}', [StudentController::class, 'details'])->middleware(['auth', 'only-admin-and-teacher']);
+Route::get('/student-add', [StudentController::class, 'create'])->middleware(['auth', 'only-admin-and-teacher']);
+Route::post('/students', [StudentController::class, 'insert'])->middleware(['auth', 'only-admin-and-teacher']);
+Route::put('student-edit/{id}', [StudentController::class, 'edit'])->middleware(['auth', 'only-admin-and-teacher']);
+Route::get('student-del/{id}', [StudentController::class, 'del'])->middleware(['auth', 'only-admin']);
+Route::delete('student-delete/{id}', [StudentController::class, 'delete'])->middleware(['auth', 'only-admin']);
+Route::get('/student-deleted', [StudentController::class, 'deleted'])->middleware(['auth', 'only-admin']);
+Route::get('/student/{id}/restore', [StudentController::class, 'restore'])->middleware(['auth', 'only-admin']);
 
 
-Route::get('/class', [ClassController::class, 'class']);
-Route::get('/class/{id}', [ClassController::class, 'details']);
-Route::get('/class/student/{id}', [StudentController::class, 'details']);
-Route::get('/class-add', [ClassController::class, 'create']);
+Route::get('/class', [ClassController::class, 'class'])->middleware('auth');
+Route::get('/class/{id}', [ClassController::class, 'details'])->middleware('auth');
+Route::get('/class/student/{id}', [StudentController::class, 'details'])->middleware('auth');
+Route::get('/class-add', [ClassController::class, 'create'])->middleware('auth');
 
-Route::get('/extracurricular', [ExtracurricularController::class, 'index']);
-Route::get('/extracurricular/{id}', [ExtracurricularController::class, 'details']);
+Route::get('/extracurricular', [ExtracurricularController::class, 'index'])->middleware('auth');
+Route::get('/extracurricular/{id}', [ExtracurricularController::class, 'details'])->middleware('auth');
 
-Route::get('/teacher', [TeacherController::class, 'index']);
-Route::get('/teacher/{id}', [TeacherController::class, 'details']);
+Route::get('/teacher', [TeacherController::class, 'index'])->middleware('auth');
+Route::get('/teacher/{id}', [TeacherController::class, 'details'])->middleware('auth');
